requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic-settings==2.6.1
PyJWT==2.8.0
cryptography==41.0.7
python-multipart==0.0.6
requests~=2.32.5
bcrypt==4.1.2 